fix(incentive): Fix fund validation on ExpandPosition
