+++ contracts/liquidity_hub/pool-network/incentive/src/execute/open_position.rs
@@ -1,8 +1,9 @@
-use cosmwasm_std::{to_binary, DepsMut, Env, MessageInfo, Response, StdError, Uint128, WasmMsg};
+use cosmwasm_std::{DepsMut, Env, MessageInfo, Response, StdError, Uint128, WasmMsg};
 use white_whale::pool_network::incentive::OpenPosition;
 
 use crate::{
     error::ContractError,
+    funds_validation::validate_funds_sent,
     state::{ADDRESS_WEIGHT, CONFIG, GLOBAL_WEIGHT, OPEN_POSITIONS},
     weight::calculate_weight,
 };
@@ -42,31 +43,7 @@ pub fn open_position(
         .unwrap_or_else(|| Ok(info.sender.clone()))?;
 
     // ensure that user gave us an allowance for the token amount
-    let allowance: cw20::AllowanceResponse = deps.querier.query_wasm_smart(
-        lp_token.clone(),
-        &cw20::Cw20QueryMsg::Allowance {
-            owner: info.sender.clone().into_string(),
-            spender: env.contract.address.clone().into_string(),
-        },
-    )?;
-
-    if allowance.allowance < amount {
-        return Err(ContractError::MissingPositionDeposit {
-            allowance_amount: allowance.allowance,
-            deposited_amount: amount,
-        });
-    }
-
-    // send the lp deposit to us
-    let messages = vec![WasmMsg::Execute {
-        contract_addr: lp_token.into_string(),
-        msg: to_binary(&cw20::Cw20ExecuteMsg::TransferFrom {
-            owner: info.sender.into_string(),
-            recipient: env.contract.address.into_string(),
-            amount,
-        })?,
-        funds: vec![],
-    }];
+    let transfer_token_msg = validate_funds_sent(&deps.as_ref(), env, lp_token, info, amount)?;
 
     // ensure an existing position with this unbonding time doesn't exist
     let existing_position = OPEN_POSITIONS
@@ -101,5 +78,5 @@ pub fn open_position(
         Ok(user_weight.unwrap_or_default().checked_add(weight)?)
     })?;
 
-    Ok(Response::new().add_messages(messages))
+    Ok(Response::new().add_message(transfer_token_msg))
 }
